% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.R
\name{align}
\alias{align}
\title{align}
\usage{
align(
  chunk_size = 15,
  step_size = 0.5,
  all_files = NULL,
  path_recordings = NULL,
  path_chunks = NULL,
  chunk_seq = NULL,
  keys_id = NULL,
  keys_rec = NULL,
  blank = 15,
  wing = 10,
  ffilter_from = NULL,
  save_pdf = FALSE,
  quiet = FALSE
)
}
\arguments{
\item{chunk_size}{numeric, duration in minutes of the chunks to output. Default is `15`.}

\item{step_size}{numeric, duration in seconds of the bins for signal compression before cross correlation.
Default is `0.5`.}

\item{all_files}{character vector, paths to all raw recordings to consider. If `NULL` files are listed
based on the argument `path_recordings`.}

\item{path_recordings}{character, the path where the raw recordings are stored. Can be nested in folders,
in this case provide the top-level folder.}

\item{path_chunks}{character, the path where aligned chunks should be stored.}

\item{chunk_seq}{numeric vector or `NULL`. If supplied only these chunks are rerun.}

\item{keys_id}{character vector of length 2. The characters before and after the unique ID of the
individual or microphone. This can be in the file name or in the folder structure. E.g., if the path to the
recording is `../data/week_1/recording_mic1.wav` the keys would be `c('recording', '.wav')`.}

\item{keys_rec}{character vector of length 2. The characters before and after the unique ID of the
recording. This can be in the file name or in the folder structure. E.g., if the path to the recording
is `../data/week_1/recording_mic1.wav` the keys would be `c('data/', '_mic')`.}

\item{blank}{numeric, the duration in minutes to be discarded at the beginning and end of the recording.}

\item{wing}{numeric, the duration in minutes to load before and after each chunk to improve alignment. This
is not saved with the aligned chunk.}

\item{ffilter_from}{numeric, frequency in Hz for the high-pass filter.}

\item{save_pdf}{logical, if `TRUE` a pdf is saved with a page per chunk that shows all the aligned
recordings.}

\item{quiet}{logical, if `TRUE` no messages are printet.}
}
\value{
saves all the aligned chunks in the location specific by `path_chunks`.
}
\description{
Aligns multiple recordings (.wav files). It assumes all microphones are within recording range
of each other most of the time.
}
