% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.detections.R
\name{export.detections}
\alias{export.detections}
\title{export.detections}
\usage{
export.detections(detections, sr = 1, path_out = "out.txt")
}
\arguments{
\item{detections}{data.frame, the object generated by `call.detect`.}

\item{sr}{numeric, the sampling rate of the wave on which detections were
run. Default to `1`, which allows users to transform the start and end times
before feeding the data.frame to this function.}

\item{path_out}{character, the path including file name where to store the
txt file. Default is `'out.txt'`.}
}
\value{
Stores a Raven Lite readable selection table.
}
\description{
Exports detection table from `call.detect` into a txt file
that can be read by Raven Lite. All columns other than those containing
start and end times are filled with 0 or ''.
}
