% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.selection.tables.R
\name{load.selection.tables}
\alias{load.selection.tables}
\title{load.selection.tables}
\usage{
load.selection.tables(path_selection_tables)
}
\arguments{
\item{path_selection_tables}{the path to the folder containing selection tables. Folder should not contain
any other files.}
}
\value{
Returns data frame with all selection tables.
}
\description{
Loads multiple Raven selection tables into one dataframe.
 Also adds a column with file-selection
}
