% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/o.to.m.R
\name{o.to.m}
\alias{o.to.m}
\title{o.to.m}
\usage{
o.to.m(o, n = seq(sqrt(length(o) + 1) + 1))
}
\arguments{
\item{o}{the vector containing the values for the lower triangular (required)}

\item{n}{the names for the rows and columns of the matrix (optional)}
}
\value{
Returns a matrix where it assumes that `m[lower.tri(m)] = o`.
}
\description{
Transforms a vector into a matrix where it assumes that the vector values are the lower
triangular of the matrix: `m[lower.tri(m)] = o`. It includes 0 on the diagonal.
}
\examples{
m = matrix(1:9, nrow = 3, ncol = 3)
o = m[lower.tri(m)]
m_new = o.to.m(o)
}
