% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple.cc.R
\name{simple.cc}
\alias{simple.cc}
\title{simple.cc}
\usage{
simple.cc(s1, s2, norm = F)
}
\arguments{
\item{dat}{A data set object}

\item{id}{Scope (e.g., country codes or individual IDs)}

\item{time}{Time (e.g., time periods are given by years, months, ...)}
}
\value{
A data frame object that contains a summary of a sample that
    can later be converted to a TeX output using \code{overview_print}
}
\description{
Simple cross correlation of two vectors. Uses zero embedding to find optimal overlap. Also
has an option to normalise by the longest vector (divides final difference by length).
This version returns the time difference for best overlap.
}
\examples{
data(toydata)
output_table <- overview_tab(dat = toydata, id = ccode, time = year)
}
