% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/better.spectro.R
\name{better.spectro}
\alias{better.spectro}
\title{better.spectro}
\usage{
better.spectro(
  wave,
  main = "",
  wl = 512,
  ovl = 450,
  xlim = "free",
  ylim = "free",
  mar = rep(3, 4)
)
}
\arguments{
\item{wave}{wave object, e.g., from `load.wave` or `readWave`.}

\item{main}{title for the spectrogram. Default is not title.}

\item{wl}{numeric, window length in samples. Default is `512`.}

\item{ovl}{numeric, overlap in samples. Default is `450`.}

\item{xlim}{numeric vector of length 2, limits for the x-axis. Default is no limits ('free').}

\item{ylim}{numeric vector of length 2, limits for the y-axis. Default is no limits ('free').}

\item{mar}{numeric vector of length 4, the margins of the plot for the `impagep` function.
Default is `rep(3, 4)`.}
}
\value{
Plots the spectrogram to current window.
}
\description{
Creates a spectrogram and plots this to the current window.
}
