% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.selection.table.R
\name{load.selection.table}
\alias{load.selection.table}
\title{load.selection.table}
\usage{
load.selection.table(path_selection_table)
}
\arguments{
\item{path_selection_table}{the path to the file containing the selection
table.}
}
\value{
Returns data frame with all selections.
}
\description{
Loads single Raven selection table into a dataframe.
}
\examples{
require(callsync)
require(seewave)
require(tuneR)
files = system.file("extdata", "", package = "callsync")
st = load.selection.tables(path_selection_tables = 'files')

}
