% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.selection.tables.audacity.R
\name{load.selection.tables.audacity}
\alias{load.selection.tables.audacity}
\title{load.selection.tables.audacity}
\usage{
load.selection.tables.audacity(path_selection_tables)
}
\arguments{
\item{path_selection_tables}{character, the path to the folder containing selection tables. Folder should
not contain any other txt files.}
}
\value{
Returns data frame with all selection tables.
}
\description{
Loads multiple Audacity selection tables into one data frame.
}
\examples{
require(callsync)
require(seewave)
require(tuneR)
files = system.file("extdata", "audacity", package = "callsync")
st = load.selection.tables.audacity(path_selection_tables = files)

}
