% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.wave.R
\name{load.wave}
\alias{load.wave}
\title{load.wave}
\usage{
load.wave(path_audio_file, from = 0, to = Inf, ffilter_from = NULL)
}
\arguments{
\item{path_audio_file}{the path to the .wav file}

\item{from}{time in seconds from where to start the loading of the audio file.
Default is `0` which loads the whole file.}

\item{to}{time in seconds until where to load the audio file. Default is `Inf` which loads the whole file.}

\item{ffilter_from}{numeric, frequency in Hz for the high-pass filter.
Default is `NULL`, which does not apply a filter.}
}
\value{
Returns an R wave object.
}
\description{
Wrapper function for `readWave` from *tuneR*.
Also optionally applies `ffilter` from *seewave*.
}
\examples{
require(callsync)
require(seewave)
require(tuneR)
file = system.file("extdata", "wave_1.wav", package = "callsync")
wave = load.wave(file)

}
