% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.trace.multiple.R
\name{measure.trace.multiple}
\alias{measure.trace.multiple}
\title{measure.trace.multiple}
\usage{
measure.trace.multiple(traces, new_waves, waves, detections, path_pdf = NULL)
}
\arguments{
\item{traces}{a list of data frames, e.g., the output of the `trace.fund` function. Should contain columns
with time = time in seconds, fund = fundamental frequency in Hz and missing = logical indicating if the
fundamental was detected (`T`) or interpolated (`F`).}

\item{new_waves}{a list of wave objects, should only contain the call.}

\item{waves}{a named list of wave objects, should not be resized. The names of the list are used for the
file name column and spectrogram titles.}

\item{detections}{the detections.}

\item{path_pdf}{numeric or `NULL`, where to store the pdf. If `NULL` no pdf is stored.}
}
\value{
Returns a data frame with all measurements.
}
\description{
Takes several measurements on multiple fundamental frequency traces.
}
