% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.trace.multiple.R
\name{measure.trace.multiple}
\alias{measure.trace.multiple}
\title{measure.trace}
\usage{
measure.trace.multiple(traces, new_waves, waves, snr = 10, path_pdf = NULL)
}
\arguments{
\item{trace}{data frame, e.g., the output of the `trace.fund` function. Should contain columns with
time = time in seconds, fund = fundamental frequency in Hz and missing = logical indicating if the
fundamental was detected (`T`) or interpolated (`F`).}

\item{sr}{sample rate of the wave object used for `trace.fund`.}
}
\value{
Returns a data frame with all measurements.
}
\description{
Takes several measurements on a fundamental frequency trace.
}
