% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.trace.multiple.R
\name{measure.trace.multiple}
\alias{measure.trace.multiple}
\title{measure.trace.multiple}
\usage{
measure.trace.multiple(
  traces,
  new_waves = NULL,
  waves = NULL,
  detections = NULL,
  sr = NULL,
  path_pdf = NULL
)
}
\arguments{
\item{traces}{a list of data frames, e.g., the output of the `trace.fund`
function. Should contain columns with time = time in seconds, fund =
fundamental frequency in Hz and missing = logical indicating if the
fundamental was detected (`TRUE`) or interpolated (`FALSE`). If the list is
named the names will be used as file names in the output.}

\item{new_waves}{a list of wave objects, should only contain the call.}

\item{waves}{a list of wave objects, should not be resized.}

\item{detections}{the detections.}

\item{sr}{numeric, sample rate of the waves objects used for the traces.
Only needed if `waves` is `NULL`.}

\item{path_pdf}{numeric or `NULL`, where to store the pdf. If `NULL` no pdf
is stored.}
}
\value{
Returns a data frame with all measurements.
}
\description{
Takes several measurements on multiple fundamental frequency
traces.
}
\examples{
require(callsync)
require(seewave)
require(tuneR)
file = system.file("extdata", "", package = "callsync")
files = list.files(file, pattern = "wave_*", full.names = TRUE)
waves = lapply(files, load.wave)
new_waves = waves
detections = lapply(waves, call.detect)
traces = lapply(waves, trace.fund)
mt = measure.trace.multiple(traces = traces, waves = waves,
                            new_waves = new_waves, detections = detections)

}
