% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure.trace.R
\name{measure.trace}
\alias{measure.trace}
\title{measure.trace}
\usage{
measure.trace(trace, sr = 44100, hop = 5)
}
\arguments{
\item{trace}{data frame, e.g., the output of the `trace.fund` function. Should contain columns with
time = time in seconds, fund = fundamental frequency in Hz and missing = logical indicating if the
fundamental was detected (`TRUE`) or interpolated (`FALSE`).}

\item{sr}{sample rate of the wave object used for `trace.fund`.}

\item{hop}{the `hop` parameter used to generate the trace.}
}
\value{
Returns a dataframe with all measurements.
}
\description{
Takes several measurements on a fundamental frequency trace.
}
