% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliding.pixel.comparison.R
\name{sliding.pixel.comparison}
\alias{sliding.pixel.comparison}
\title{sliding.pixel.comparison}
\usage{
sliding.pixel.comparison(s1, s2, step_size = 1)
}
\arguments{
\item{s1}{numeric matrix, the first spectrogram.}

\item{s2}{numeric matrix, the second spectrogram.}

\item{step_size}{numeric, how many pixels should be moved for each step. Default is `1`.}
}
\value{
Returns the distance at the point of maximal signal overlap.
}
\description{
Can be used to run spectrographic cross correlation. Both spectrograms are zero-padded
and slid over each other. For each step the difference is computed. The function returns the absolute
difference at the point at the minimum (maximal signal overlap).
}
